+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 - Go to Definition shows both symbol and task for clocks.
 - Improved documentation support for actions with multiple overloads.
+- Improved documentation for several actions.
 
 ### Fixed